perf(tools): lazy-load orchestrator implementations
