+++ src/engine/worker-runtime-capabilities.ts
@@ -1,0 +1,5 @@
+/**
+ * Native worker registries expose no nested Clio dispatch contract. This is a
+ * compile-time runtime capability, not a setting.
+ */
+export const WORKER_RUNTIME_MEDIATES_CLIO_DISPATCH = false;
