+++ src/tools/monitor.ts
@@ -1,7 +1,5 @@
 import { readFileSync } from "node:fs";
-import { Type } from "typebox";
 import { sleep } from "../core/timers.js";
-import { ToolNames } from "../core/tool-names.js";
 import { renderAgentLedgerBoard } from "../domains/dispatch/agent-ledger-store.js";
 import type { DurableAssignmentRecord } from "../domains/dispatch/assignment-store.js";
 import type { DispatchContract } from "../domains/dispatch/contract.js";
@@ -14,8 +12,8 @@ import {
 	type RunReceiptVerification,
 } from "../domains/dispatch/types.js";
 import { COST_NOT_MEASURED, costAggregateForAmount, formatCostAggregate } from "../domains/observability/index.js";
-import { StringEnum } from "../engine/ai.js";
 import type { DispatchRunEventRegistry } from "./dispatch.js";
+import { monitorToolSurface } from "./monitor-surface.js";
 import type { ToolInvokeOptions, ToolResult, ToolSpec } from "./registry.js";
 import { truncateUtf8 } from "./truncate-utf8.js";
 import { receiptEvidenceLabels, workerTextLabel, workerTextNonEvidenceNotices } from "./worker-evidence.js";
@@ -707,35 +705,7 @@ async function runCollect(
 
 export function createMonitorTool(deps: MonitorToolDeps): ToolSpec {
 	return {
-		name: ToolNames.Monitor,
-		description:
-			"Inspect known dispatched runs. Parent-model mid-run observation requires detach:true because ordinary dispatch auto-waits. wait observes without collecting. collect is the authoritative terminal batch operation; collect detached runs before final synthesis. receipt exposes stored evidence. Receipt integrity, evidence verification, briefing provenance, and project-context provenance are separate fields.",
-		parameters: Type.Object({
-			run_id: Type.Optional(
-				Type.String({ description: "Run id from dispatch output or monitor list; omit with mode=list." }),
-			),
-			mode: Type.Optional(
-				StringEnum(["status", "peek", "receipt", "list", "wait", "collect", "tools"], {
-					description:
-						"What to return. Defaults to status when run_id is present and list when it is absent. status, peek, receipt, tools, and wait each observe one run and require a run_id; list takes none. tools answers what a run executed: its tool calls with outcomes, plus per-tool totals from the receipt.",
-				}),
-			),
-			batch_id: Type.Optional(Type.String({ description: "Detached batch id from dispatch detach:true (mode=collect)." })),
-			run_ids: Type.Optional(
-				Type.Array(Type.String(), {
-					description:
-						"Explicit run ids for mode=collect; a one-element array is also accepted by single-run modes when run_id is absent.",
-				}),
-			),
-			timeout_ms: Type.Optional(
-				Type.Number({
-					description:
-						"mode=wait: max ms to block (default 60000, capped at 600000); mode=collect never blocks and ignores this value with a notice.",
-				}),
-			),
-		}),
-		baseActionClass: "read",
-		executionMode: "parallel",
+		...monitorToolSurface,
 		async run(args, options): Promise<ToolResult> {
 			const explicitRunId = typeof args.run_id === "string" ? args.run_id.trim() : "";
 			const rawRunIds = Array.isArray(args.run_ids) ? args.run_ids : null;
