+++ src/tools/steer.ts
@@ -1,8 +1,6 @@
-import { Type } from "typebox";
-import { ToolNames } from "../core/tool-names.js";
 import type { DispatchContract } from "../domains/dispatch/contract.js";
-import { StringEnum } from "../engine/ai.js";
 import type { ToolResult, ToolSpec } from "./registry.js";
+import { steerToolSurface } from "./steer-surface.js";
 
 /**
  * The steer tool: control a running dispatched worker. action=guide injects a
@@ -62,16 +60,7 @@ function cancel(deps: SteerToolDeps, runId: string): ToolResult {
 
 export function createSteerTool(deps: SteerToolDeps): ToolSpec {
 	return {
-		name: ToolNames.Steer,
-		description:
-			"Control a running native worker whose run id is already available. Parent-model mid-run control requires detach:true because ordinary dispatch auto-waits and dispatch/steer are sequential; the interactive operator/TUI may steer an active synchronous native run through the dispatch contract. ACP runs have no input channel.",
-		parameters: Type.Object({
-			run_id: Type.String({ description: "Run id from dispatch output or monitor list." }),
-			action: StringEnum(["guide", "cancel"], { description: "What to do." }),
-			message: Type.Optional(Type.String({ description: "action=guide: the steering message." })),
-		}),
-		baseActionClass: "dispatch",
-		executionMode: "sequential",
+		...steerToolSurface,
 		async run(args): Promise<ToolResult> {
 			const runId = typeof args.run_id === "string" ? args.run_id.trim() : "";
 			if (runId.length === 0) return { kind: "error", message: "steer: missing run_id argument" };
